separate MainMenu component
